// src/shared/enum/setting.enum.ts
export enum SettingEnum {
  SUPPORT_CONTACT_NUMBER = 'SUPPORT_CONTACT_NUMBER',
  SUPPORT_CONTACT_EMAIL = 'SUPPORT_CONTACT_EMAIL',
  EMAIL_ENABLE = 'EMAIL_ENABLE',
  SENDGRID_API_KEY = 'SENDGRID_API_KEY',
  SENDER_EMAIL = 'SENDER_EMAIL',
  MAINTENANCE_MODE = 'MAINTENANCE_MODE',
  ENABLE_SMS = 'ENABLE_SMS',
  MESSAGE_SERVICE_SID = 'MESSAGE_SERVICE_SID',
  LITE_SCREENING_LOCKED_PERIOD = "LITE_SCREENING_LOCKED_PERIOD"
}

export enum AppPlatform {
  ANDROID = 'md',
  IOS = 'ios',
  APP_GALLERY = 'app-gallery',
}